updateresponse

+++ src/main/java/com/schoolpayment/team/dto/request/StudentRequest.java
@@ -1,6 +1,7 @@
 package com.schoolpayment.team.dto.request;
 
 
+import com.schoolpayment.team.model.ClassEntity;
 import lombok.Data;
 
 import java.time.LocalDate;
@@ -9,11 +10,12 @@ import java.time.LocalDate;
 public class StudentRequest {
     private String nis;
     private String name;
-    private int classId;
+    private Long classId;
     private String address;
     private String phoneNumber;
 
     public LocalDate getBirthdate() {
         return LocalDate.now();
     }
+
 }
